Expose watchdog flags used in Stackless
These flags are not really needed as we don't aim to support different
setups. However to keep this part of the work separate from implementation
they have so far been preserved.

This keeps it simple for possible others that may come onto the project
to follow without needing context.

However, comments have been added to the relevant function which states this
information.