Add c-api tests for all api functions. Change how cextension is imported. Implement watchdog
Various small changes required for c-api tests to pass correctly

Watchdog now accepts timeout, run n tasklets functionality kept as useful for tests

Including scheduler still requires a base package 'scheduler'
If blue is successfully imported then extension is imported through blue supporting build flavor
If blue is not imported then it will search for an environment variable for BUILDFLAVOR
The latter is useful for when using scheduler with tests.
However, the necessity of having another packaged named simply 'scheduler' is more complex:
If the cextension is imported directly '_scheduler' then the capsule name also needs to match
eg. '_scheduler' would extend with debug flavor to '_scheduler_debug'
The capsule name defined in the scheduler public header is 'scheduler._C_API'
This will now fail to import as there is no module 'scheduler'
It should be 'scheduler_debug._C_API' to work correctly
If we wanted this to work without a 'scheduler' package we would need to generate a different header per flavor
which is messy when consuming the header.
Therefore having the intermediate package in there makes all this work without snowflake changes to blue.