Refactor to schedule manager cleanup memory, Add sendthrow in capi
Schedule managers should now garbage collect once they are no longer used
They will be created on demand and tasklets created on the same thread will
keep them alive. When all tasklets are destroyed and no other references to
the schedule manager are held they will be collected.

PyScheduler_CallMethod_Main has been retired as a dedicated function for this
no longer makes sense. In theory one can just call scheduler, however in practice
it makes more sense to get a reference to the main scheduler, then call main loop,
then once done it can release the main scheduler which should destroy it if no
memory problems are present.

Method makes it simple to track some memory issues, this check has been added to tests
some of which are now failing. I have not looked closely at them but all failing tests
will need attention.

Added PyChannel_SendThrow and the related python code as this is now a requirement of
Carbon-io. It was previously not in the initial requirement stub for api