Handle possible memory leaks when losing reference to channels, fix greenlet arg leak
If a blocked channel reference goes out of scope it is not cleaned up automatically
due to the fact that the stack currently has a reference to the channel in the cstate
this needs to be cleaned up before it can be destroyed.

This should never happen in an ideal world however it is possible that this could occur
so must be handled.

In order to do this a weak list of channels is stored on no module level and functions
for querying and clearing all the channels has been supplied.

Also fix was made to not leak when passing arguments to greenlets on switch. Arguments
should only be passed during the first switch. Subsequent switches caused a leak.