Add module level channel tracking and ability to clear all channels
It is possible to lose accessible references to channel which have blocked tasklets
this means that those tasklets will never unblock until interpreter teardown.
change adds a channel mem check to tests and fix to an exposed channel leak.
A module level function has been added to force all channels to clear.
This functionality needs testing in relation to threads.