Remove auto channel unblock code which relies on incorrect assertion
This code was supposed to unblock a channel that you may have lost a ref for
It is only applicable in a certain situation, not all and the assertion is now
in practice incorrect. Removing it also makes the code cleaner.

We will have to just have channels with lost accessible references unblock when
module finishes which should occur.