Fix bug when closing channel closing too soon.
If the channel was blocked on send it could get marked as closed before
the last tasklet had been unblocked from the channel.

This was due to the place where the channel close status check was. It
was being checked each time the balance was changed. However during internal
channel transfer the balance gets incremented and decremented before it
ends up at the correct value that the user will see. This up and down
sometimes was causing the balance to be momentarily 0 which would then
signal to the channel that it can be closed. However before control got
back to the user the balance would be correctly incremented by 1 to reflect
the true state. Therefore the channel has something blocked on it but had
incorrectly already been marked as closed.

This was not caught by tests as they didn't cover the blocked on send and only
blocked on receive. Another good example why a symmetrical channel design would
be safer.