Fix crash in tasklet.remove. Fix bind + others not returning tasklets.
Tasklets removed using tasklet.remove were not being unmarked scheduled.
This lead to a crash when operations were made on them such as run.
This is fixed + test coverage added

Many tasklet operations should return the tasklet. Such as bind and insert
these have been added + test coverage added

Cross thread operations have been blocked off with error messages. This
behaviour is different from Stackless but as the change is non trivial
and likely this behaviour is not required I have added the errors so we
are at least aware if it happens in the client. Previously calling these
operations across thread would probably lead to a crash.