Fix crash and leak possible with channels and kill pending.
With channel with preference sender completes a transfer but is pending killed
before the receiver has resumed the code path would crash.

With channel with preference receiver completes a transfer but is pending killed
before the sender has resumed the code path would result in the receiving tasklet
being leaked. This would also cause a leak for the schedule manager.

Tests have been added to cover both cases.