Remove usage of build flavor mechanism
Blue depends on scheduler, and subsequently supporting build flavors creates a circular dependency that causes import issues for blue down the line. Therefore, stop using build flavors, much like it was done for carbon-io, which is also a dependency for blue. The alternative would be to move both components into blue itself, but that sounds strange, too.