Fix build error on CI
This time for real. As it turns out, the problem was not the target-dependent generator expression, but rather, the use of a multi-config generator on TeamCity. In that scenario one must have config-specific output files for `file(GENERATE ...)`. Alas, instead of trying to forward the configuration value to the C++ sources and detecting the correct header name to include, this now ensures that a header with the correct name gets copied from the configuration specific header.