Fix a tasklet leak caused by calling PyTasklet_Setup through C API
PyTasklet_Setup does not return a PyObject* but rather a boolean, so it does not make sense that it should increase the reference count.

Calling the method twice would set the reschedule flag on a tasklet, which would result in it not getting killed right away. Since this does not seem like desired behavior to me, I also updated the Kill method to remove the reschedule flag if present.

This led me to modify `test_kill_pending_true` to expect a runcount of 1 after a pending kill followed by running the tasklet, which seems to be in line with what happens in Stackless Python.

This change should to fix a leak in Blue when sending events through PostEvent as is done with `Destiny::OnExitWarp` and `Destiny::OnActivatingWarp`